requirements.txt
streamlit>=1.24.0
supabase>=1.0.0
pandas
python-dotenv>=0.21.0
requests>=2.28.0
plotly>=5.13.0
tqdm>=4.65.0
numpy 